Rework infer path structure
